Corrected mistake in oracle logic.
We were using a topological sort of the lattice representation DAG as a way of advancing the search, but that was flawed. Corrected by keeping track of 'valid' and 'invalid' nodes -- i.e., those that belong or not to the sublattice of disjoint successors of an ideal X -- with the oracles redefined accordingly.